requirements.txt
alembic==1.8.0
autopep8==1.6.0
click==8.1.3
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
protobuf==4.21.2
pycodestyle==2.8.0
python-dotenv==0.20.0
SQLAlchemy==1.4.39
toml==0.10.2
Werkzeug==2.1.2
